# Performance: Elide async/await in JobHost when we're initialized