Part of #2794. This avoids the async/await overhead when we know we're initialized in the JobHost itself, saving one async/await transition between `CallAsync` and `CallAsyncCore` in the common case. Overall: a little bit of perf from avoiding work here with a tab bit more code, but justifiable given the hot path, IMO.

I got some perf run numbers here to help - 1-2% throughput gains eliminating the state machine overhead here:

#### Baseline
```lua
➜ .\bombard.exe -k -l -c 100 -d 300s "http://localhost:5000/api/HttpTriggerAnon?name=bob"
Bombarding http://localhost:5000/api/HttpTriggerAnon?name=bob for 5m0s using 100 connection(s)
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     76740.15   14985.40   96909.99
  Latency        1.30ms   297.99us   154.89ms
  Latency Distribution
     50%     1.00ms
     75%     1.01ms
     90%     2.00ms
     95%     2.00ms
     99%     4.00ms
  HTTP codes:
    1xx - 0, 2xx - 23025315, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    18.37MB/s
➜ .\bombard.exe -k -l -c 100 -d 300s "http://localhost:5000/api/HttpTriggerAnon?name=bob"
Bombarding http://localhost:5000/api/HttpTriggerAnon?name=bob for 5m0s using 100 connection(s)
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     76805.44   14731.65   96499.04
  Latency        1.30ms   262.78us   106.84ms
  Latency Distribution
     50%     1.00ms
     75%     1.01ms
     90%     2.00ms
     95%     2.00ms
     99%     4.00ms
  HTTP codes:
    1xx - 0, 2xx - 23037660, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    18.38MB/s
```

#### After (with async elide)
```lua
➜ .\bombard.exe -k -l -c 100 -d 300s "http://localhost:5000/api/HttpTriggerAnon?name=bob"
Bombarding http://localhost:5000/api/HttpTriggerAnon?name=bob for 5m0s using 100 connection(s)
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     78226.60   15508.83   99323.60
  Latency        1.27ms   281.40us    79.07ms
  Latency Distribution
     50%     1.00ms
     75%     1.00ms
     90%     2.00ms
     95%     2.00ms
     99%     4.00ms
  HTTP codes:
    1xx - 0, 2xx - 23469246, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    18.73MB/s
➜ .\bombard.exe -k -l -c 100 -d 300s "http://localhost:5000/api/HttpTriggerAnon?name=bob"
Bombarding http://localhost:5000/api/HttpTriggerAnon?name=bob for 5m0s using 100 connection(s)
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     78157.14   15314.57   99508.22
  Latency        1.28ms   259.43us    80.05ms
  Latency Distribution
     50%     1.00ms
     75%     1.01ms
     90%     2.00ms
     95%     2.00ms
     99%     4.00ms
  HTTP codes:
    1xx - 0, 2xx - 23442864, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    18.71MB/s
```