@NickCraver 2022-01-25
@safihamid This is the one I was mentioning in the call - eliminating the async/await here can shave a tad off cold start as well.